GUI/cure/requirements.txt
alembic==1.0.1
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.1
Flask==1.0.2
Flask-API==1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.7
ItsDangerous==1.0.0
Jinja2==2.10
json2html==1.2.1
ldap3==2.5.1
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
pkg-resources==0.0.0
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
python-coveralls==2.9.1
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
PyYAML==3.13
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24
Werkzeug==0.14.1
WTForms==2.2.1
